Windows: export roken allocators
Use rk_alloc, rk_calloc, rk_free, rk_realloc instead of Windows
C RTL allocators.  The OpenAFS source tree has a bad habit of
allocating memory in one module and freeing it in another.  This
is not a problem for POSIX but is disaster on Windows.  This change
ensures that all OpenAFS modules share the same allocator.

Change-Id: Ia57f1ba351e8dc24e7fbba87bbc9595bcf001ed6
Reviewed-on: http://gerrit.openafs.org/7625
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>