Windows: cm_DiscardSCache volumeCreationDate
commit f716962ab41847af4450d0a361f5de9195b32ed0
clears the cm_scache_t volumeCreationDate field.
It shouldn't because the volumeCreationDate is not a property of
the callback.  It is a property of the status information which
does not change simply because the callback expires.

Reviewed-on: http://gerrit.openafs.org/7678
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Tested-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit 64a37fa9f916cd1e8f0ec9bb252fda13aa8b4cd4)

Change-Id: I18be397abb0cd223ac508c6addd11d00c64c54ba
Reviewed-on: http://gerrit.openafs.org/7712
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>