FBSD: don't reference libc_r; no release OpenAFS works on still uses it
libc_r is the old user-mode threading library.  Modern versions of
FreeBSD don't include it, and the conditionals here that (nearly
always) override it with the correct library, libpthread, are true
on every version of FreeBSD for which OpenAFS might plausibly be
compiled.  So just use the correct library all the time.

Change-Id: I8f813814973cec3173f44636523c3bf0bee27b3e
Reviewed-on: http://gerrit.openafs.org/7764
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementix.org>